Refactoring of Task aliases to make the approach to Tasks more unified accross the whole code. (#108)
I have replace all aliases onto the name `Task` with just the bare Elm
`Task x a`. So no more `IO a`, or `Install.Task a` or other specialized
versions of Task.

`Reporting.Task` was unnecessarily complicated by using `Result`, when
`Task` already has space in its constructors for managing errors.

All `Task` monad helper code was moved into `Utils.Task.Extra`.

By doing this, it is much more obvious to work with Tasks, and see how
they flow accross the whole project in a more unified way. Tasks are
central to this project as it does a lot of IO. It is now simpler to do
things like writing a debug IO task for example, that can be used in any
Task context within the project.